fix: drop fragile chart modifiers, wrap CGFloat for SplitsTable bar
- HRZoneBars / PaceZoneBars: remove chartXAxis modifier entirely. The
  default time-axis is informative for horizontal bars and avoids any
  chartXAxis(.hidden) / empty-closure overload mismatch.
- SplitsTable: wrap paceBarRatio (Double) in CGFloat() to match the
  pattern used elsewhere (PaceTable.swift) and avoid CGFloat * Double
  inference issues.

+++ RunJourney/Features/Detail/Components/SplitsTable.swift
@@ -104,7 +104,7 @@ struct SplitsTable: View {
                     Capsule()
                         .fill(PaceUtils.paceBarColor(pace: paceInt, basePace: averagePaceSec))
                         .frame(
-                            width: geo.size.width * PaceUtils.paceBarRatio(pace: paceInt, basePace: averagePaceSec),
+                            width: geo.size.width * CGFloat(PaceUtils.paceBarRatio(pace: paceInt, basePace: averagePaceSec)),
                             height: 6
                         )
                 }
